# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: NexBank Digital Banking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "200":
          description: Email verified
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Reset email sent
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset

  # ── ACCOUNTS ──────────────────────────────────────────────────────────
...
  /beneficiaries:
    get:
      operationId: listBeneficiaries
      tags: [beneficiaries]
      summary: List saved beneficiaries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Beneficiaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeneficiaryListResponse"
    post:
      operationId: createBeneficiary
      tags: [beneficiaries]
      summary: Add beneficiary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeneficiaryInput"
      responses:
        "201":
          description: Beneficiary added
  /beneficiaries/{id}:
    patch:
      operationId: updateBeneficiary
      tags: [beneficiaries]
      summary: Update beneficiary
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeneficiaryUpdate"
      responses:
        "200":
          description: Updated
    delete:
      operationId: deleteBeneficiary
      tags: [beneficiaries]
      summary: Delete beneficiary
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked read
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── PROFILE ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          {}
      required:
        - success

    # ── Auth Schemas ──────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        isEmailVerified:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        kycStatus:
          type: string
        createdAt:
          type: string
      required: [id, firstName, lastName, email, role]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            accessToken:
              type: string
            refreshToken:
              type: string
      required: [success]
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string

    # ── Account Schemas ───────────────────────────────────────────────
...
    Beneficiary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: ["string", "null"]
        bankName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        nickname:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, accountNumber, bankName]
    BeneficiaryInput:
      type: object
      required: [name, accountNumber, bankName]
      properties:
        name:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: ["string", "null"]
        bankName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        nickname:
          type: ["string", "null"]
    BeneficiaryUpdate:
      type: object
      properties:
        name:
          type: string
        bankName:
          type: string
        nickname:
          type: ["string", "null"]
        isFavorite:
          type: boolean
    BeneficiaryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Beneficiary"

    # ── Notification Schemas ──────────────────────────────────────────
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [TRANSACTION, SECURITY, SYSTEM, PROMOTION, ALERT]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, message, isRead]
    NotificationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"

    # ── Profile Schemas ───────────────────────────────────────────────
...