# lib/api-zod/src/generated/types/beneficiaryListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NexBank Digital Banking API
 * OpenAPI spec version: 1.0.0
 */
import type { Beneficiary } from './beneficiary';

export interface BeneficiaryListResponse {
  success?: boolean;
  data?: Beneficiary[];
}