# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NexBank Digital Banking API
 * OpenAPI spec version: 1.0.0
 */

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  isEmailVerified?: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  kycStatus?: string;
  createdAt?: string;
}